chore(system): extract the executor to its own module
