+++ cli/build.gradle
@@ -34,6 +34,7 @@ dependencies {
     implementation project(":storage-local")
 
     // Kestra server components
+    implementation project(":executor")
     implementation project(":scheduler")
     implementation project(":webserver")
     implementation project(":worker")
