+++ executor/src/test/resources/allure.properties
@@ -1,0 +1,1 @@
+allure.results.directory=build/allure-results
