+++ jdbc-h2/build.gradle
@@ -5,6 +5,7 @@ configurations {
 dependencies {
     implementation project(":core")
     implementation project(":jdbc")
+    implementation project(":executor")
 
     implementation("io.micronaut.sql:micronaut-jooq")
     runtimeOnly("com.h2database:h2")
