+++ jdbc-postgres/build.gradle
@@ -5,6 +5,7 @@ configurations {
 dependencies {
     implementation project(":core")
     implementation project(":jdbc")
+    implementation project(":executor")
 
     implementation("io.micronaut.sql:micronaut-jooq")
     runtimeOnly("org.postgresql:postgresql")
